% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/aeep_mcmc.R
\docType{package}
\name{aeep}
\alias{aeep}
\alias{aeep-package}
\title{aeep: Estimate Bayesian Models in Parallel}
\description{
This packages provides tools to estimate bayesian models in parallel and more.
}
\section{Foo functions}{

The foo functions ...
}

