% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/comb_par.R
\name{comb_par}
\alias{comb_par}
\title{Parametric Combination of Posteriors}
\usage{
comb_par(post_list)
}
\arguments{
\item{post_list}{A \code{list} of matrices containing samples from the
sub-posteriors}
}
\value{
A list containing a vector of means and variances of the full
posterior
}
\description{
Parametric Combination of Posteriors
}

