% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{combine_np}
\alias{combine_np}
\title{Non-parametric Combination of Sub-posteriors}
\usage{
combine_np(post_list)
}
\arguments{
\item{post_list}{A list containing the sub-posterior samples, stored in matrices}
}
\value{
A matrix containing the combined posterior samples
}
\description{
Non-parametric Combination of Sub-posteriors
}

