% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/combine_p.R
\name{combine_p}
\alias{combine_p}
\title{Parametric Combination of Posteriors}
\usage{
combine_p(post_list)
}
\arguments{
\item{post_list}{A list of matrices containing samples from the
sub-posteriors}
}
\value{
A list containing a vector of means and the variance-covariance matrix
of the full posterior
}
\description{
Parametric Combination of Posteriors
}

