% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{combine_sp}
\alias{combine_sp}
\title{Semi-parametric Combination of Sub-posteriors}
\usage{
combine_sp(post_list)
}
\arguments{
\item{post_list}{A list containing the sub-posterior samples, stored in matrices}
}
\value{
A matrix containing the samples from the sub-posterior product function
}
\description{
Semi-parametric Combination of Sub-posteriors
}

