% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/irtpar.R
\docType{package}
\name{foo}
\alias{foo}
\alias{foo-package}
\title{irtpar: A Package for Estimating Bayesian IRT Models in Parallel}
\description{
The foo package provides three categories of important functions:
foo, bar and baz.
}
\section{Foo functions}{

The foo functions ...
}

