% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/irtpar.R
\docType{package}
\name{irtpar}
\alias{irtpar}
\alias{irtpar-package}
\title{irtpar: A Package for Estimating Bayesian IRT Models in Parallel}
\description{
This packages provides tools to estimate bayesian models in parallel and more.
}
\section{Foo functions}{

The foo functions ...
}

