% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/par2pl.R
\name{par2pl}
\alias{par2pl}
\title{Fit 2pl IRT Models in Parallel}
\usage{
par2pl(data, cores = NULL, iter, warmup = floor(iter/2), chains,
  combine = "parametric")
}
\arguments{
\item{data}{A matrix or data.frame containing the complete data
set.}

\item{cores}{The number of cores to be used to fit the model. This implies the
number of partitions created. If not supplied the number of cores is detected
automatically.}

\item{iter}{Number of iterations for each mcmc sampler.}

\item{warmup}{Number of warmup samples for each sampler.}

\item{chains}{Number of chains to run on each sub-sample.}

\item{combine}{Method to be used to combine sub-posteriors to the full posterior
(\code{"parametric"}, \code{"semi-parametric"} (or \code{"non-parametric"}))}
}
\value{
A list containing the full posterior and the sub-posteriors.

If \code{combine == "parametric"} The vector of means and the variance-covariance
matrix of the full-posterior is returned

If \code{combine == "non-parametric"} The full posterior is returned as a
matrix with number of rows equal to number of iterations minus warmup and
number of columns equal to the number of free parameters in the model.

Note: For now the function only returns the item parameters. INdividual level
parameters will be implemented shortly.
}
\description{
Fit 2pl IRT Models in Parallel
}

