% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/parallel_mcmc.R
\name{parallel_mcmc}
\alias{parallel_mcmc}
\title{Fit Bayesian Models in Parallel}
\usage{
parallel_mcmc(data, cores, combine = "parametric", fit)
}
\arguments{
\item{data}{A \code{matrix} or \code{data.frame} containing the complete data
set.}

\item{cores}{The number of cores to be used to fit the model. This implies the
number of partitions created. If not supplied the number of cores is detected
automatically.}

\item{combine}{Method to be used to combine sub-posteriors to the full posterior}

\item{fit}{A function fitting the model to the data subsample, first argument
must be data. No other arguments allowed (for now)
(\code{"parametric"}, \code{"semi-parametric"} (or \code{"non-parametric"}))}
}
\value{
A list containing the full posterior and the sub-posteriors
}
\description{
Fit Bayesian Models in Parallel
}

