% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/sim_irt.R
\name{sim_irt}
\alias{sim_irt}
\title{Simulate IRT Data}
\usage{
sim_irt(n_subjects, n_items, p_theta = c(0, 1), p_alpha = c(-1, 0.3),
  p_beta = c(0, 1), seed = NULL, params = FALSE)
}
\arguments{
\item{n_subjects}{Number of subjects (rows).}

\item{n_items}{Number of items (columns).}

\item{p_theta}{A vector of length 2 containing mean and standard deviation for
the normal distribution the thetas are generated from.}

\item{p_alpha}{A vector of length 2 containing mean and standard deviation for
the log-normal distribution the alphas are generated from.}

\item{p_beta}{A vector of length 2 containing mean and standard deviation for
the normal distribution the alphas are generated from.}

\item{seed}{A seed for the random number generator.}

\item{params}{a logical value indicating wether true parameters should be
returned}
}
\value{
If \code{params == TRUE} A list containing the generated data
set and the true parameter vectors.

If \code{params == FALSE} a matrix
containing the generated data
}
\description{
Simulate IRT Data
}
\details{
Generates data according to
\deqn{Y \sim Bernoulli(\pi_{jk})}
}

