% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/parallel_mcmc.R
\name{parallel_mcmc}
\alias{parallel_mcmc}
\title{Fit Bayesian Models in Parallel}
\usage{
parallel_mcmc(data, cores, combine = NULL, fun)
}
\arguments{
\item{data}{A matrix or data.frame containing the complete data
set.}

\item{cores}{The number of cores to be used to fit the model. This implies the
number of partitions created. If not supplied the number of cores is detected
automatically.}

\item{combine}{Method to be used to combine sub-posteriors to the full posterior
(\code{"parametric"}, \code{"semi-parametric"} (or \code{"non-parametric"}))}

\item{fun}{A function fitting the model to the data subsample. For now the
function must only have two arguments. The first beeing the data, the second
beeing the number of partitions.}
}
\value{
A list containing the full posterior and the sub-posteriors
}
\description{
Fit Bayesian Models in Parallel
}

